requirements.txt
anthropic
chromadb
pdfplumber
python-docx
playwright
pyyaml
python-dotenv
